requirements.txt
bs4
requests
azure-storage-blob
googlemaps